# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sewa Barang API Documentation",
        "description": "Dokumentasi API untuk platform dashboard Sewa Barang",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/regions": {
            "get": {
                "tags": [
                    "Regions"
                ],
                "summary": "Daftar wilayah",
                "description": "Mengambil semua daftar wilayah/lokasi",
                "operationId": "getRegionList",
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil data"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Regions"
                ],
                "summary": "Buat wilayah baru",
                "description": "Menambahkan wilayah baru ke dalam database",
                "operationId": "storeRegion",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Jakarta Selatan"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Wilayah berhasil dibuat"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/regions-paginated": {
            "get": {
                "tags": [
                    "Regions"
                ],
                "summary": "Daftar wilayah (Paginated)",
                "description": "Mengambil daftar wilayah dengan sistem pagination",
                "operationId": "getRegionPaginated",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Jumlah data per halaman",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil data"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/regions/{id}": {
            "put": {
                "tags": [
                    "Regions"
                ],
                "summary": "Update wilayah",
                "description": "Memperbarui data wilayah berdasarkan ID",
                "operationId": "updateRegion",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Wilayah",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Jakarta Pusat"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Wilayah berhasil diperbarui"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Regions"
                ],
                "summary": "Hapus wilayah",
                "description": "Menghapus wilayah berdasarkan ID",
                "operationId": "destroyRegion",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Wilayah",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Wilayah berhasil dihapus"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}